Initial commit on hateoas repo, update Newtonsoft.Json
These files have been moved over from the feature23/odata-lite
repository where they were originally created. This new repo is
dedicated exclusively to the F23.Hateoas library now.

This commit, in addition to bringing over the files, updates
Newtonsoft.Json to fix a vulnerability, and prepares the project for the
1.1.0 release.

This will be the last release with Newtonsoft.Json included. 2.0.0 will
replace its usage with System.Text.Json, and a new package will be added
for Newtonsoft.Json compatibility.